requirements.txt
python-telegram-bot[job-queue]==20.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
apscheduler==3.10.4